refactor(authz): split lib.rs into per-concern modules
The whole crate was one 1484-line `lib.rs` — the at-cut view, the
subgroup-tree walk, credential admission, the decision itself and all 11
tests. Split along the seams it already had, with no behavior change: the
public API is byte-identical, since every item is re-exported flat from
`lib.rs` and the modules stay private.

- `authorize.rs`   the decision: `authorize`, `required_mask_for`, and the
                   device-binding precondition every op but a link passes
- `view.rs`        `AclView` + the binding types, and the predicates that
                   answer from a single map lookup
- `inheritance.rs` a second `impl AclView`: the three questions that walk
                   the subgroup tree, together so their differences read as
                   deliberate rather than as drift
- `admission.rs`   `admit_device_link` / `fold_device_link` /
                   `admit_key_rotation` — the rules `calimero-projection`
                   folds against, stated once so they cannot diverge
- `error.rs`       `Rejected`

Tests move out of the source files into one `src/tests/` tree, one file per
module under test plus shared fixtures in `src/tests/support.rs`.

Beyond the move, three fixes the split surfaced:

- two rustdoc warnings that predate this change (an unresolved link to
  `calimero_projection`, and `may` linking the private `DEFAULT_MEMBER_MASK`)
  — the crate is now doc-warning-clean;
- two doc comments that had drifted off `bind_test_devices` and
  `bind_account` onto an unrelated test are back on their functions, minus a
  sentence promising device-precondition tests this crate does not have;
- AGENTS.md's public-API table listed three crate-private constants as API
  and omitted the account-plane functions `calimero-projection` calls.

Same 11 tests, all passing; `cargo check --workspace --all-targets`, clippy
`-D warnings`, `cargo fmt --check` and `cargo doc` are clean.
